+++ ts-sdk/client/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @crypticdot/fusionamm-client
 
+## 1.0.59
+
+## 1.0.58
+
 ## 1.0.57
 
 ## 1.0.56